Move style related tests to format
