+++ pandas/tests/io/test_excel.py
@@ -17,7 +17,6 @@ import numpy as np
 
 import pandas as pd
 from pandas import DataFrame, Index, MultiIndex
-from pandas.io.formats.excel import ExcelFormatter
 from pandas.io.parsers import read_csv
 from pandas.io.excel import (
     ExcelFile, ExcelWriter, read_excel, _XlwtWriter, _Openpyxl1Writer,
@@ -2351,219 +2350,3 @@ class ExcelWriterEngineTests(object):
                 check_called(
                     lambda: df.to_excel(
                         'something.xls', engine='dummy'))
-
-
-@pytest.fixture
-def styled_dataframe():
-    def style(df):
-        # XXX: RGB colors not supported in xlwt
-        return DataFrame([['font-weight: bold', '', ''],
-                          ['', 'color: blue', ''],
-                          ['', '', 'text-decoration: underline'],
-                          ['border-style: solid', '', ''],
-                          ['', 'font-style: italic', ''],
-                          ['', '', 'text-align: right'],
-                          ['background-color: red', '', ''],
-                          ['', '', ''],
-                          ['', '', ''],
-                          ['', '', '']],
-                         index=df.index, columns=df.columns)
-
-    pytest.importorskip('jinja2')
-    df = DataFrame(np.random.randn(10, 3))
-    return df.style.apply(style, axis=None)
-
-
-def assert_equal_style(cell1, cell2):
-    # XXX: should find a better way to check equality
-    # Neither OpenPyXl's Cell, nor its style objects have __eq__ defined
-    assert cell1.alignment.__dict__ == cell2.alignment.__dict__
-    assert cell1.border.__dict__ == cell2.border.__dict__
-    assert cell1.fill.__dict__ == cell2.fill.__dict__
-    assert cell1.font.__dict__ == cell2.font.__dict__
-    assert cell1.number_format == cell2.number_format
-    assert cell1.protection.__dict__ == cell2.protection.__dict__
-
-
-@pytest.mark.parametrize('engine', [
-    pytest.mark.xfail('xlwt', reason='xlwt does not support '
-                                     'openpyxl-compatible style dicts'),
-    'xlsxwriter',
-    'openpyxl',
-])
-def test_styler_to_excel(engine, styled_dataframe):
-    def custom_converter(css):
-        # use bold iff there is custom style attached to the cell
-        if css.strip(' \n;'):
-            return {'font': {'bold': True}}
-        return {}
-
-    pytest.importorskip('cssdecl')
-    pytest.importorskip(engine)
-
-    if engine == 'openpyxl' and openpyxl_compat.is_compat(major_ver=1):
-        pytest.xfail('openpyxl1 does not support some openpyxl2-compatible '
-                     'style dicts')
-
-    styled = styled_dataframe
-    df = styled.data
-
-    # Prepare spreadsheets
-
-    with ensure_clean('.xlsx' if engine != 'xlwt' else '.xls') as path:
-        writer = ExcelWriter(path, engine=engine)
-        df.to_excel(writer, sheet_name='frame')
-        df.style.to_excel(writer, sheet_name='unstyled')
-        styled.to_excel(writer, sheet_name='styled')
-        ExcelFormatter(styled, style_converter=custom_converter).write(
-            writer, sheet_name='custom')
-
-    # For engines other than openpyxl 2, we only smoke test
-    if engine != 'openpyxl':
-        return
-    if not openpyxl_compat.is_compat(major_ver=2):
-        pytest.skip('incompatible openpyxl version')
-
-    # (1) compare DataFrame.to_excel and Styler.to_excel when unstyled
-    n_cells = 0
-    for col1, col2 in zip(writer.sheets['frame'].columns,
-                          writer.sheets['unstyled'].columns):
-        assert len(col1) == len(col2)
-        for cell1, cell2 in zip(col1, col2):
-            assert cell1.value == cell2.value
-            assert_equal_style(cell1, cell2)
-            n_cells += 1
-
-    # ensure iteration actually happened:
-    assert n_cells == (10 + 1) * (3 + 1)
-
-    # (2) check styling with default converter
-    n_cells = 0
-    for col1, col2 in zip(writer.sheets['frame'].columns,
-                          writer.sheets['styled'].columns):
-        assert len(col1) == len(col2)
-        for cell1, cell2 in zip(col1, col2):
-            ref = '%s%d' % (cell2.column, cell2.row)
-            # XXX: this isn't as strong a test as ideal; we should
-            #      differences are exclusive
-            if ref == 'B2':
-                assert not cell1.font.bold
-                assert cell2.font.bold
-            elif ref == 'C3':
-                assert cell1.font.color.rgb != cell2.font.color.rgb
-                assert cell2.font.color.rgb == '000000FF'
-            elif ref == 'D4':
-                assert cell1.font.underline != cell2.font.underline
-                assert cell2.font.underline == 'single'
-            elif ref == 'B5':
-                assert not cell1.border.left.style
-                assert (cell2.border.top.style ==
-                        cell2.border.right.style ==
-                        cell2.border.bottom.style ==
-                        cell2.border.left.style ==
-                        'medium')
-            elif ref == 'C6':
-                assert not cell1.font.italic
-                assert cell2.font.italic
-            elif ref == 'D7':
-                assert (cell1.alignment.horizontal !=
-                        cell2.alignment.horizontal)
-                assert cell2.alignment.horizontal == 'right'
-            elif ref == 'B8':
-                assert cell1.fill.fgColor.rgb != cell2.fill.fgColor.rgb
-                assert cell1.fill.patternType != cell2.fill.patternType
-                assert cell2.fill.fgColor.rgb == '00FF0000'
-                assert cell2.fill.patternType == 'solid'
-            else:
-                assert_equal_style(cell1, cell2)
-
-            assert cell1.value == cell2.value
-            n_cells += 1
-
-    assert n_cells == (10 + 1) * (3 + 1)
-
-    # (3) check styling with custom converter
-    n_cells = 0
-    for col1, col2 in zip(writer.sheets['frame'].columns,
-                          writer.sheets['custom'].columns):
-        assert len(col1) == len(col2)
-        for cell1, cell2 in zip(col1, col2):
-            ref = '%s%d' % (cell2.column, cell2.row)
-            if ref in ('B2', 'C3', 'D4', 'B5', 'C6', 'D7', 'B8'):
-                assert not cell1.font.bold
-                assert cell2.font.bold
-            else:
-                assert_equal_style(cell1, cell2)
-
-            assert cell1.value == cell2.value
-            n_cells += 1
-
-    assert n_cells == (10 + 1) * (3 + 1)
-
-
-@pytest.mark.parametrize('engine', [
-    pytest.mark.xfail('xlwt', reason='xlwt does not support '
-                                     'openpyxl-compatible style dicts'),
-    'xlsxwriter',
-    'openpyxl',
-])
-def test_styler_to_excel_no_cssdecl(engine, styled_dataframe):
-    def custom_converter(css):
-        # use bold iff there is custom style attached to the cell
-        if css.strip(' \n;'):
-            return {'font': {'bold': True}}
-        return {}
-
-    try:
-        import cssdecl  # noqa
-    except ImportError:
-        with pytest.raises(ImportError) as rec:
-            pd.DataFrame({"A": [1, 2]}).style.to_excel(engine)
-        assert rec.match("not installed")
-    else:
-        pytest.skip('Test only run if cssdecl not installed')
-
-    pytest.importorskip(engine)
-
-    styled = styled_dataframe
-    df = styled.data
-
-    # Prepare spreadsheets
-
-    with ensure_clean('.xlsx' if engine != 'xlwt' else '.xls') as path:
-        writer = ExcelWriter(path, engine=engine)
-        df.to_excel(writer, sheet_name='frame')
-        with pytest.raises(ImportError):
-            # default style_converter requires cssdecl
-            df.style.to_excel(writer, sheet_name='unstyled')
-            styled.to_excel(writer, sheet_name='styled')
-
-        ExcelFormatter(styled, style_converter=custom_converter).write(
-            writer, sheet_name='custom')
-
-        writer.save()
-
-        openpyxl = pytest.importorskip('openpyxl')
-        if openpyxl_compat.is_compat(major_ver=1):
-            # smoke test only
-            return
-
-        wb = openpyxl.load_workbook(path)
-
-        # check styling with custom converter
-        n_cells = 0
-        for col1, col2 in zip(wb['frame'].columns,
-                              wb['custom'].columns):
-            assert len(col1) == len(col2)
-            for cell1, cell2 in zip(col1, col2):
-                ref = '%s%d' % (cell2.column, cell2.row)
-                if ref in ('B2', 'C3', 'D4', 'B5', 'C6', 'D7', 'B8'):
-                    assert not cell1.font.bold
-                    assert cell2.font.bold
-                else:
-                    assert_equal_style(cell1, cell2)
-
-                assert cell1.value == cell2.value
-                n_cells += 1
-
-        assert n_cells == (10 + 1) * (3 + 1)
